requirements.txt
aws-cdk-lib==2.102.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha==2.31.1a0
python-decouple==3.6
